+++ swap/src/database.rs
@@ -2,10 +2,13 @@ pub use alice::Alice;
 pub use bob::Bob;
 
 use anyhow::{anyhow, bail, Context, Result};
+use itertools::Itertools;
+use libp2p::PeerId;
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use std::path::Path;
+use std::str::FromStr;
 use uuid::Uuid;
 
 mod alice;
@@ -38,16 +41,34 @@ impl Display for Swap {
     }
 }
 
+#[derive(thiserror::Error, Debug, Clone, Copy, PartialEq)]
+#[error("Not in the role of Alice")]
+struct NotAlice;
+
+#[derive(thiserror::Error, Debug, Clone, Copy, PartialEq)]
+#[error("Not in the role of Bob")]
+struct NotBob;
+
 impl Swap {
+    pub fn try_into_alice(self) -> Result<Alice> {
+        match self {
+            Swap::Alice(alice) => Ok(alice),
+            Swap::Bob(_) => bail!(NotAlice),
+        }
+    }
+
     pub fn try_into_bob(self) -> Result<Bob> {
         match self {
             Swap::Bob(bob) => Ok(bob),
-            Swap::Alice(_) => bail!("Swap instance is not Bob"),
+            Swap::Alice(_) => bail!(NotBob),
         }
     }
 }
 
-pub struct Database(sled::Db);
+pub struct Database {
+    swaps: sled::Tree,
+    peers: sled::Tree,
+}
 
 impl Database {
     pub fn open(path: &Path) -> Result<Self> {
@@ -56,22 +77,51 @@ impl Database {
         let db =
             sled::open(path).with_context(|| format!("Could not open the DB at {:?}", path))?;
 
-        Ok(Database(db))
+        let swaps = db.open_tree("swaps")?;
+        let peers = db.open_tree("peers")?;
+
+        Ok(Database { swaps, peers })
+    }
+
+    pub async fn insert_peer_id(&self, swap_id: Uuid, peer_id: PeerId) -> Result<()> {
+        let peer_id_str = peer_id.to_string();
+
+        let key = serialize(&swap_id)?;
+        let value = serialize(&peer_id_str).context("Could not serialize peer-id")?;
+
+        self.peers.insert(key, value)?;
+
+        self.peers
+            .flush_async()
+            .await
+            .map(|_| ())
+            .context("Could not flush db")
+    }
+
+    pub fn get_peer_id(&self, swap_id: Uuid) -> Result<PeerId> {
+        let key = serialize(&swap_id)?;
+
+        let encoded = self
+            .peers
+            .get(&key)?
+            .ok_or_else(|| anyhow!("No peer-id found for swap id {} in database", swap_id))?;
+
+        let peer_id: String = deserialize(&encoded).context("Could not deserialize peer-id")?;
+        Ok(PeerId::from_str(peer_id.as_str())?)
     }
 
     pub async fn insert_latest_state(&self, swap_id: Uuid, state: Swap) -> Result<()> {
         let key = serialize(&swap_id)?;
         let new_value = serialize(&state).context("Could not serialize new state value")?;
 
-        let old_value = self.0.get(&key)?;
+        let old_value = self.swaps.get(&key)?;
 
-        self.0
+        self.swaps
             .compare_and_swap(key, old_value, Some(new_value))
             .context("Could not write in the DB")?
             .context("Stored swap somehow changed, aborting saving")?;
 
-        // TODO: see if this can be done through sled config
-        self.0
+        self.swaps
             .flush_async()
             .await
             .map(|_| ())
@@ -82,7 +132,7 @@ impl Database {
         let key = serialize(&swap_id)?;
 
         let encoded = self
-            .0
+            .swaps
             .get(&key)?
             .ok_or_else(|| anyhow!("Swap with id {} not found in database", swap_id))?;
 
@@ -90,22 +140,42 @@ impl Database {
         Ok(state)
     }
 
-    pub fn all(&self) -> Result<Vec<(Uuid, Swap)>> {
-        self.0
-            .iter()
-            .map(|item| match item {
-                Ok((key, value)) => {
-                    let swap_id = deserialize::<Uuid>(&key);
-                    let swap = deserialize::<Swap>(&value).context("Failed to deserialize swap");
+    pub fn all_alice(&self) -> Result<Vec<(Uuid, Alice)>> {
+        self.all_alice_iter().collect()
+    }
 
-                    match (swap_id, swap) {
-                        (Ok(swap_id), Ok(swap)) => Ok((swap_id, swap)),
-                        (Ok(_), Err(err)) => Err(err),
-                        _ => bail!("Failed to deserialize swap"),
-                    }
-                }
-                Err(err) => Err(err).context("Failed to retrieve swap from DB"),
-            })
+    fn all_alice_iter(&self) -> impl Iterator<Item = Result<(Uuid, Alice)>> {
+        self.all_swaps_iter().map(|item| {
+            let (swap_id, swap) = item?;
+            Ok((swap_id, swap.try_into_alice()?))
+        })
+    }
+
+    pub fn all_bob(&self) -> Result<Vec<(Uuid, Bob)>> {
+        self.all_bob_iter().collect()
+    }
+
+    fn all_bob_iter(&self) -> impl Iterator<Item = Result<(Uuid, Bob)>> {
+        self.all_swaps_iter().map(|item| {
+            let (swap_id, swap) = item?;
+            Ok((swap_id, swap.try_into_bob()?))
+        })
+    }
+
+    fn all_swaps_iter(&self) -> impl Iterator<Item = Result<(Uuid, Swap)>> {
+        self.swaps.iter().map(|item| {
+            let (key, value) = item.context("Failed to retrieve swap from DB")?;
+
+            let swap_id = deserialize::<Uuid>(&key)?;
+            let swap = deserialize::<Swap>(&value).context("Failed to deserialize swap")?;
+
+            Ok((swap_id, swap))
+        })
+    }
+
+    pub fn unfinished_alice(&self) -> Result<Vec<(Uuid, Alice)>> {
+        self.all_alice_iter()
+            .filter_ok(|(_swap_id, alice)| !matches!(alice, Alice::Done(_)))
             .collect()
     }
 }
@@ -187,26 +257,106 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn can_fetch_all_keys() {
+    async fn all_swaps_as_alice() {
         let db_dir = tempfile::tempdir().unwrap();
         let db = Database::open(db_dir.path()).unwrap();
 
-        let state_1 = Swap::Alice(Alice::Done(AliceEndState::BtcPunished));
-        let swap_id_1 = Uuid::new_v4();
-        db.insert_latest_state(swap_id_1, state_1.clone())
+        let alice_state = Alice::Done(AliceEndState::BtcPunished);
+        let alice_swap = Swap::Alice(alice_state.clone());
+        let alice_swap_id = Uuid::new_v4();
+        db.insert_latest_state(alice_swap_id, alice_swap)
             .await
-            .expect("Failed to save second state");
+            .expect("Failed to save alice state 1");
 
-        let state_2 = Swap::Bob(Bob::Done(BobEndState::SafelyAborted));
-        let swap_id_2 = Uuid::new_v4();
-        db.insert_latest_state(swap_id_2, state_2.clone())
+        let alice_swaps = db.all_alice().unwrap();
+        assert_eq!(alice_swaps.len(), 1);
+        assert!(alice_swaps.contains(&(alice_swap_id, alice_state)));
+
+        let bob_state = Bob::Done(BobEndState::SafelyAborted);
+        let bob_swap = Swap::Bob(bob_state);
+        let bob_swap_id = Uuid::new_v4();
+        db.insert_latest_state(bob_swap_id, bob_swap)
             .await
-            .expect("Failed to save first state");
+            .expect("Failed to save bob state 1");
 
-        let swaps = db.all().unwrap();
+        let err = db.all_alice().unwrap_err();
 
-        assert_eq!(swaps.len(), 2);
-        assert!(swaps.contains(&(swap_id_1, state_1)));
-        assert!(swaps.contains(&(swap_id_2, state_2)));
+        assert_eq!(err.downcast_ref::<NotAlice>().unwrap(), &NotAlice);
+    }
+
+    #[tokio::test]
+    async fn all_swaps_as_bob() {
+        let db_dir = tempfile::tempdir().unwrap();
+        let db = Database::open(db_dir.path()).unwrap();
+
+        let bob_state = Bob::Done(BobEndState::SafelyAborted);
+        let bob_swap = Swap::Bob(bob_state.clone());
+        let bob_swap_id = Uuid::new_v4();
+        db.insert_latest_state(bob_swap_id, bob_swap)
+            .await
+            .expect("Failed to save bob state 1");
+
+        let bob_swaps = db.all_bob().unwrap();
+        assert_eq!(bob_swaps.len(), 1);
+        assert!(bob_swaps.contains(&(bob_swap_id, bob_state)));
+
+        let alice_state = Alice::Done(AliceEndState::BtcPunished);
+        let alice_swap = Swap::Alice(alice_state);
+        let alice_swap_id = Uuid::new_v4();
+        db.insert_latest_state(alice_swap_id, alice_swap)
+            .await
+            .expect("Failed to save alice state 1");
+
+        let err = db.all_bob().unwrap_err();
+
+        assert_eq!(err.downcast_ref::<NotBob>().unwrap(), &NotBob);
+    }
+
+    #[tokio::test]
+    async fn can_save_swap_state_and_peer_id_with_same_swap_id() -> Result<()> {
+        let db_dir = tempfile::tempdir().unwrap();
+        let db = Database::open(db_dir.path()).unwrap();
+
+        let alice_id = Uuid::new_v4();
+        let alice_state = Alice::Done(AliceEndState::BtcPunished);
+        let alice_swap = Swap::Alice(alice_state);
+        let peer_id = PeerId::random();
+
+        db.insert_latest_state(alice_id, alice_swap.clone()).await?;
+        db.insert_peer_id(alice_id, peer_id).await?;
+
+        let loaded_swap = db.get_state(alice_id)?;
+        let loaded_peer_id = db.get_peer_id(alice_id)?;
+
+        assert_eq!(alice_swap, loaded_swap);
+        assert_eq!(peer_id, loaded_peer_id);
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_reopen_db() -> Result<()> {
+        let db_dir = tempfile::tempdir().unwrap();
+        let alice_id = Uuid::new_v4();
+        let alice_state = Alice::Done(AliceEndState::BtcPunished);
+        let alice_swap = Swap::Alice(alice_state);
+
+        let peer_id = PeerId::random();
+
+        {
+            let db = Database::open(db_dir.path()).unwrap();
+            db.insert_latest_state(alice_id, alice_swap.clone()).await?;
+            db.insert_peer_id(alice_id, peer_id).await?;
+        }
+
+        let db = Database::open(db_dir.path()).unwrap();
+
+        let loaded_swap = db.get_state(alice_id)?;
+        let loaded_peer_id = db.get_peer_id(alice_id)?;
+
+        assert_eq!(alice_swap, loaded_swap);
+        assert_eq!(peer_id, loaded_peer_id);
+
+        Ok(())
     }
 }